% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dp_to_dapps.R
\name{convert_dp_to_dapps}
\alias{convert_dp_to_dapps}
\title{Wrapper function that converts a `.dp` file into a DAPPSR projection output
list format.}
\usage{
convert_dp_to_dapps(dp_fpath)
}
\arguments{
\item{dp_fpath}{A file path string to the `.dp` input file}
}
\value{
a list of five keys: "data", "summary_measures", "age_summary", "warnings", and "error". The "data" key contains a list of six data frames for each year of the projection:
male and female population by age, male and female births by age of mother, male and female deaths by age, male and female migration by age,
male and female life tables by age. Years correspond to the mid-year. The "summary_measures" key contains a set of summary tables about the vital events of the
projection. The "age_summary" key contains a set of summary data by age group and by year. The "warnings" and "error" keys contain a vector of warnings encountered
or the error that caused the projection process to stop. In the event of an error, the "data" key should contain partial results.
}
\description{
Wrapper function that converts a `.dp` file into a DAPPSR projection output
list format.
}
