% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_extract_functions.R
\name{dp_findstartindex}
\alias{dp_findstartindex}
\title{Given a tag, finds the first index marking the start of the tagged data}
\usage{
dp_findstartindex(dp, tag, tagcol = 1)
}
\arguments{
\item{dp}{A DemoProj data frame}

\item{tag}{A tag string}

\item{tagcol}{The column number of dp to search for the tag. Defaults to `1`.}
}
\value{
an integer
}
\description{
Given a tag, finds the first index marking the start of the tagged data
}
