% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_extract_functions.R
\name{read_dp}
\alias{read_dp}
\title{Read in .dp or .pjn files as data frames}
\usage{
read_dp(fpath)
}
\arguments{
\item{fpath}{A file path string to the `.dp` input file}
}
\value{
a data frame
}
\description{
Read in .dp or .pjn files as data frames
}
