% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_dp}
\alias{read_dp}
\title{read_dp}
\usage{
read_dp(fpath)
}
\arguments{
\item{fpath}{A file path string to the `.dp` input file.}
}
\value{
A data frame.
}
\description{
Read in .dp or .pjn files as data frames
}
\author{
Vania Wang
}
